perf(agents): project catalog runtime capabilities once
