+++ src/agents/prepared-model-runtime.configured-catalog.ts
@@ -6,13 +6,11 @@ import type { ModelCatalogSnapshot } from "./model-catalog.types.js";
 import {
   toStaticCatalogEntry,
   type PreparedConfiguredRuntimeModel,
-  type PreparedRuntimeCapabilityModel,
 } from "./prepared-model-runtime.configured.js";
 import type { ModelRegistry } from "./sessions/model-registry.js";
 
 type ConfiguredCatalogAgentFacts = {
   configuredModelRefs: readonly ModelCatalogRef[];
-  runtimeCapabilityModels: readonly PreparedRuntimeCapabilityModel[];
 };
 
 type ConfiguredCatalogWorkspaceFacts = {
@@ -57,57 +55,14 @@ function createConfiguredModelCatalogSnapshot(params: {
     }
   }
   const configuredEntries = [...entries.values()];
-  const materializedEntries = materializeRuntimeCapabilities(
-    configuredEntries,
-    params.agentFacts.runtimeCapabilityModels,
-  );
-  const staticEntries = materializeRuntimeCapabilities(
-    params.configuredRuntimeModels.map(({ model }) => toStaticCatalogEntry(model)),
-    params.agentFacts.runtimeCapabilityModels,
+  const staticEntries = params.configuredRuntimeModels.map(({ model }) =>
+    toStaticCatalogEntry(model),
   );
   return {
-    entries: materializedEntries,
-    routeVariants: materializedEntries,
+    entries: configuredEntries,
+    routeVariants: configuredEntries,
     ...(staticEntries.length > 0 ? { staticEntries } : {}),
   };
-}
-
-/**
- * Configured views omit runtime-only rows. Retain the concrete route's
- * capabilities on the logical row so downstream projections do not rediscover
- * or depend on an absent runtime sibling.
- */
-export function materializeRuntimeCapabilities(
-  entries: readonly ModelCatalogEntry[],
-  runtimeCapabilityModels: readonly PreparedRuntimeCapabilityModel[],
-): ModelCatalogEntry[] {
-  const runtimeByKey = new Map(
-    runtimeCapabilityModels.map(({ provider, modelId, model }) => [
-      modelCatalogEntryKey({ provider, id: modelId }),
-      toStaticCatalogEntry(model),
-    ]),
-  );
-  return entries.map((entry) => {
-    const runtime = runtimeByKey.get(modelCatalogEntryKey(entry));
-    if (!runtime) {
-      return entry;
-    }
-    const thinkingPolicyProvider = runtime.provider;
-    if (entry.configuredReasoning !== undefined) {
-      return { ...entry, thinkingPolicyProvider };
-    }
-    const params =
-      runtime.params || entry.params ? { ...runtime.params, ...entry.params } : undefined;
-    const compat =
-      runtime.compat || entry.compat ? { ...runtime.compat, ...entry.compat } : undefined;
-    return {
-      ...entry,
-      thinkingPolicyProvider,
-      ...(runtime.reasoning !== undefined ? { reasoning: runtime.reasoning } : {}),
-      ...(params ? { params } : {}),
-      ...(compat ? { compat } : {}),
-    };
-  });
 }
 
 export function prepareConfiguredRuntimeFacts(params: {
